Resolve conflicts: keep app/auth/* as canonical, remove duplicate app/blueprints/auth/routes.py, register blueprints once, fix agent package, re-id security migration
