+++ app/agent/__init__.py
@@ -1,3 +1,5 @@
 """Agent utilities blueprint package."""
 
-__all__ = ["env_audit"]
+from .routes import agent_bp  # para import corto: from app.agent import agent_bp
+
+__all__ = ["agent_bp"]
