+++ app/agent/routes.py
@@ -10,11 +10,9 @@ from .env_audit import env_audit
 agent_bp = Blueprint("agent", __name__, url_prefix="/agent")
 
 
-@agent_bp.get("/env-audit")
+@agent_bp.route("/env-audit")
 @login_required
 def env_a():
-    """Return the environment audit report as JSON."""
-
     return jsonify(env_audit())
