+++ app/blueprints/auth/__init__.py
@@ -1,5 +1,5 @@
 from __future__ import annotations
 
-from .routes import bp_auth
+from app.auth.routes import auth_bp as bp_auth
 
 __all__ = ["bp_auth"]
